create validator utility
